% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_interface.R
\name{Drivers}
\alias{Drivers}
\title{Drivers R6 Class}
\description{
R6 wrapper for external forcing data (drivers)
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{ptr}}{External pointer to the underlying C++ Drivers object.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Drivers-initialize}{\code{Drivers$new()}}
    \item \href{#method-Drivers-set_constant}{\code{Drivers$set_constant()}}
    \item \href{#method-Drivers-set_variable}{\code{Drivers$set_variable()}}
    \item \href{#method-Drivers-set_extrapolate}{\code{Drivers$set_extrapolate()}}
    \item \href{#method-Drivers-evaluate}{\code{Drivers$evaluate()}}
    \item \href{#method-Drivers-evaluate_range}{\code{Drivers$evaluate_range()}}
    \item \href{#method-Drivers-get_names}{\code{Drivers$get_names()}}
    \item \href{#method-Drivers-clear}{\code{Drivers$clear()}}
    \item \href{#method-Drivers-clone}{\code{Drivers$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-initialize}{}}}
\subsection{\code{Drivers$new()}}{
  Initialize a `Drivers` instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-set_constant"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-set_constant}{}}}
\subsection{\code{Drivers$set_constant()}}{
  Set a constant driver value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$set_constant(name, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Driver variable name.}
      \item{\code{value}}{Constant driver value.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-set_variable"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-set_variable}{}}}
\subsection{\code{Drivers$set_variable()}}{
  Set a variable driver from paired vectors.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$set_variable(name, x, y)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Driver variable name.}
      \item{\code{x}}{Numeric vector of input coordinates (e.g., time).}
      \item{\code{y}}{Numeric vector of values corresponding to `x`.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-set_extrapolate"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-set_extrapolate}{}}}
\subsection{\code{Drivers$set_extrapolate()}}{
  Configure extrapolation behavior for a driver.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$set_extrapolate(name, extrapolate)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Driver variable name.}
      \item{\code{extrapolate}}{Logical flag controlling extrapolation beyond data range.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-evaluate}{}}}
\subsection{\code{Drivers$evaluate()}}{
  Evaluate a driver at one input value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$evaluate(name, x)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Driver variable name.}
      \item{\code{x}}{Numeric vector of input coordinates (e.g., time).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-evaluate_range"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-evaluate_range}{}}}
\subsection{\code{Drivers$evaluate_range()}}{
  Evaluate a driver over a vector of input values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$evaluate_range(name, x)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{Driver variable name.}
      \item{\code{x}}{Numeric vector of input coordinates (e.g., time).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-get_names"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-get_names}{}}}
\subsection{\code{Drivers$get_names()}}{
  Return names of configured drivers.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$get_names()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-clear}{}}}
\subsection{\code{Drivers$clear()}}{
  Remove all configured drivers.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$clear()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Drivers-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Drivers-clone}{}}}
\subsection{\code{Drivers$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Drivers$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
