% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz-interface.R
\name{LorenzSystem}
\alias{LorenzSystem}
\title{Lorenz System R6 Class}
\description{
R6 wrapper for Lorenz system
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{ptr}}{External pointer to the underlying C++ Lorenz system object.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LorenzSystem-initialize}{\code{LorenzSystem$new()}}
    \item \href{#method-LorenzSystem-pars}{\code{LorenzSystem$pars()}}
    \item \href{#method-LorenzSystem-set_params}{\code{LorenzSystem$set_params()}}
    \item \href{#method-LorenzSystem-set_state}{\code{LorenzSystem$set_state()}}
    \item \href{#method-LorenzSystem-state}{\code{LorenzSystem$state()}}
    \item \href{#method-LorenzSystem-rates}{\code{LorenzSystem$rates()}}
    \item \href{#method-LorenzSystem-set_initial_state}{\code{LorenzSystem$set_initial_state()}}
    \item \href{#method-LorenzSystem-reset}{\code{LorenzSystem$reset()}}
    \item \href{#method-LorenzSystem-clone}{\code{LorenzSystem$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-initialize}{}}}
\subsection{\code{LorenzSystem$new()}}{
  Initialize a Lorenz system object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$new(sigma, R, b)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{sigma}}{Lorenz parameter sigma.}
      \item{\code{R}}{Lorenz parameter R.}
      \item{\code{b}}{Lorenz parameter b.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-pars"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-pars}{}}}
\subsection{\code{LorenzSystem$pars()}}{
  Return current system parameters.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$pars()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-set_params}{}}}
\subsection{\code{LorenzSystem$set_params()}}{
  Set model parameters.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$set_params(params)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{params}}{Numeric vector of system parameters.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-set_state"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-set_state}{}}}
\subsection{\code{LorenzSystem$set_state()}}{
  Set system state and time.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$set_state(y, time = 0)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{y}}{Numeric state vector.}
      \item{\code{time}}{Scalar time value.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-state"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-state}{}}}
\subsection{\code{LorenzSystem$state()}}{
  Return current state.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$state()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-rates"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-rates}{}}}
\subsection{\code{LorenzSystem$rates()}}{
  Return current rates.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$rates()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-set_initial_state"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-set_initial_state}{}}}
\subsection{\code{LorenzSystem$set_initial_state()}}{
  Set initial state and initial time.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$set_initial_state(y, t0 = 0)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{y}}{Numeric state vector.}
      \item{\code{t0}}{Initial time value.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-reset"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-reset}{}}}
\subsection{\code{LorenzSystem$reset()}}{
  Reset the system to its initial condition.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$reset()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-clone}{}}}
\subsection{\code{LorenzSystem$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LorenzSystem$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
