% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_interface.R
\name{OdeControl}
\alias{OdeControl}
\title{OdeControl R6 Class}
\description{
R6 wrapper for ODE solver control settings
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{ptr}}{External pointer to the underlying C++ OdeControl object.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-OdeControl-initialize}{\code{OdeControl$new()}}
    \item \href{#method-OdeControl-get_controls}{\code{OdeControl$get_controls()}}
    \item \href{#method-OdeControl-set_controls}{\code{OdeControl$set_controls()}}
    \item \href{#method-OdeControl-set_tol_abs}{\code{OdeControl$set_tol_abs()}}
    \item \href{#method-OdeControl-set_tol_rel}{\code{OdeControl$set_tol_rel()}}
    \item \href{#method-OdeControl-set_a_y}{\code{OdeControl$set_a_y()}}
    \item \href{#method-OdeControl-set_a_dydt}{\code{OdeControl$set_a_dydt()}}
    \item \href{#method-OdeControl-set_step_size_min}{\code{OdeControl$set_step_size_min()}}
    \item \href{#method-OdeControl-set_step_size_max}{\code{OdeControl$set_step_size_max()}}
    \item \href{#method-OdeControl-set_controller}{\code{OdeControl$set_controller()}}
    \item \href{#method-OdeControl-get_controller}{\code{OdeControl$get_controller()}}
    \item \href{#method-OdeControl-set_step_size_initial}{\code{OdeControl$set_step_size_initial()}}
    \item \href{#method-OdeControl-clone}{\code{OdeControl$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-initialize}{}}}
\subsection{\code{OdeControl$new()}}{
  Initialize an `OdeControl` instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-get_controls"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-get_controls}{}}}
\subsection{\code{OdeControl$get_controls()}}{
  Get current solver control settings.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$get_controls()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_controls"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_controls}{}}}
\subsection{\code{OdeControl$set_controls()}}{
  Set all solver control settings at once.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_controls(
  tol_abs,
  tol_rel,
  a_y,
  a_dydt,
  step_size_min,
  step_size_max,
  step_size_initial
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tol_abs}}{Absolute tolerance.}
      \item{\code{tol_rel}}{Relative tolerance.}
      \item{\code{a_y}}{State scaling coefficient.}
      \item{\code{a_dydt}}{Derivative scaling coefficient.}
      \item{\code{step_size_min}}{Minimum allowed step size.}
      \item{\code{step_size_max}}{Maximum allowed step size.}
      \item{\code{step_size_initial}}{Initial step size.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_tol_abs"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_tol_abs}{}}}
\subsection{\code{OdeControl$set_tol_abs()}}{
  Set absolute tolerance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_tol_abs(tol_abs)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tol_abs}}{Absolute tolerance.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_tol_rel"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_tol_rel}{}}}
\subsection{\code{OdeControl$set_tol_rel()}}{
  Set relative tolerance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_tol_rel(tol_rel)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tol_rel}}{Relative tolerance.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_a_y"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_a_y}{}}}
\subsection{\code{OdeControl$set_a_y()}}{
  Set state scaling coefficient.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_a_y(a_y)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{a_y}}{State scaling coefficient.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_a_dydt"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_a_dydt}{}}}
\subsection{\code{OdeControl$set_a_dydt()}}{
  Set derivative scaling coefficient.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_a_dydt(a_dydt)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{a_dydt}}{Derivative scaling coefficient.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_step_size_min"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_step_size_min}{}}}
\subsection{\code{OdeControl$set_step_size_min()}}{
  Set minimum step size.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_step_size_min(step_size_min)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{step_size_min}}{Minimum allowed step size.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_step_size_max"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_step_size_max}{}}}
\subsection{\code{OdeControl$set_step_size_max()}}{
  Set maximum step size.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_step_size_max(step_size_max)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{step_size_max}}{Maximum allowed step size.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_controller"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_controller}{}}}
\subsection{\code{OdeControl$set_controller()}}{
  Set the step-size rule, `"gsl"` or `"hairer"`.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_controller(controller)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{controller}}{Which rule turns the error estimate into the next step:
`"gsl"` (the default; accept below 1.1, grow only below 0.5, otherwise
keep the step) or `"hairer"` (accept at 1, rescale after every step, as
`deSolve` and Hairer's codes do). The default is kept for continuity of
step sequences; `"hairer"` makes far fewer rejected attempts and lets the
step track a smoothing solution (odelia#64).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-get_controller"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-get_controller}{}}}
\subsection{\code{OdeControl$get_controller()}}{
  The step-size rule in force.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$get_controller()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-set_step_size_initial"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-set_step_size_initial}{}}}
\subsection{\code{OdeControl$set_step_size_initial()}}{
  Set initial step size.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$set_step_size_initial(step_size_initial)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{step_size_initial}}{Initial step size.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeControl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OdeControl-clone}{}}}
\subsection{\code{OdeControl$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OdeControl$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
