\name{odestepper-package}
\alias{odestepper-package}
\alias{odestepper}
\docType{package}
\title{\packageTitle{odestepper}}
\description{\packageDescription{odestepper}}
\details{The DESCRIPTION file: \packageDESCRIPTION{odestepper}}
\section{Package Content}{\packageIndices{odestepper}}
\author{\packageAuthor{odestepper}}
\section{Maintainer}{\packageMaintainer{odestepper}}
\references{http://www.boost.org/}
\keyword{ package }
