% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz-interface.R
\name{Lorenz_Solver}
\alias{Lorenz_Solver}
\title{Lorenz Solver R6 Class}
\description{
R6 wrapper for Lorenz ODE solver with optional AD support
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lorenz_Solver-new}{\code{Lorenz_Solver$new()}}
\item \href{#method-Lorenz_Solver-time}{\code{Lorenz_Solver$time()}}
\item \href{#method-Lorenz_Solver-state}{\code{Lorenz_Solver$state()}}
\item \href{#method-Lorenz_Solver-set_state}{\code{Lorenz_Solver$set_state()}}
\item \href{#method-Lorenz_Solver-times}{\code{Lorenz_Solver$times()}}
\item \href{#method-Lorenz_Solver-advance_adaptive}{\code{Lorenz_Solver$advance_adaptive()}}
\item \href{#method-Lorenz_Solver-advance_fixed}{\code{Lorenz_Solver$advance_fixed()}}
\item \href{#method-Lorenz_Solver-step}{\code{Lorenz_Solver$step()}}
\item \href{#method-Lorenz_Solver-reset}{\code{Lorenz_Solver$reset()}}
\item \href{#method-Lorenz_Solver-collect}{\code{Lorenz_Solver$collect()}}
\item \href{#method-Lorenz_Solver-history_size}{\code{Lorenz_Solver$history_size()}}
\item \href{#method-Lorenz_Solver-history_step}{\code{Lorenz_Solver$history_step()}}
\item \href{#method-Lorenz_Solver-history}{\code{Lorenz_Solver$history()}}
\item \href{#method-Lorenz_Solver-set_target}{\code{Lorenz_Solver$set_target()}}
\item \href{#method-Lorenz_Solver-fit}{\code{Lorenz_Solver$fit()}}
\item \href{#method-Lorenz_Solver-clone}{\code{Lorenz_Solver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$new(System_xp, control_xp, active = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-time"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-time}{}}}
\subsection{Method \code{time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-state"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-state}{}}}
\subsection{Method \code{state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-set_state"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-set_state}{}}}
\subsection{Method \code{set_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$set_state(y, time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-times"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-times}{}}}
\subsection{Method \code{times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$times()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-advance_adaptive"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-advance_adaptive}{}}}
\subsection{Method \code{advance_adaptive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$advance_adaptive(times)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-advance_fixed"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-advance_fixed}{}}}
\subsection{Method \code{advance_fixed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$advance_fixed(times)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-step"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-step}{}}}
\subsection{Method \code{step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$step()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-collect"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$collect(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-history_size"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-history_size}{}}}
\subsection{Method \code{history_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$history_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-history_step"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-history_step}{}}}
\subsection{Method \code{history_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$history_step(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-history"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-history}{}}}
\subsection{Method \code{history()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$history()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-set_target"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-set_target}{}}}
\subsection{Method \code{set_target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$set_target(times, target, obs_indices)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$fit(ic = NULL, params = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lorenz_Solver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lorenz_Solver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lorenz_Solver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
