% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz-interface.R
\name{LorenzSystem}
\alias{LorenzSystem}
\title{Lorenz System R6 Class}
\description{
R6 wrapper for Lorenz system
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LorenzSystem-new}{\code{LorenzSystem$new()}}
\item \href{#method-LorenzSystem-pars}{\code{LorenzSystem$pars()}}
\item \href{#method-LorenzSystem-set_params}{\code{LorenzSystem$set_params()}}
\item \href{#method-LorenzSystem-set_state}{\code{LorenzSystem$set_state()}}
\item \href{#method-LorenzSystem-state}{\code{LorenzSystem$state()}}
\item \href{#method-LorenzSystem-rates}{\code{LorenzSystem$rates()}}
\item \href{#method-LorenzSystem-set_initial_state}{\code{LorenzSystem$set_initial_state()}}
\item \href{#method-LorenzSystem-reset}{\code{LorenzSystem$reset()}}
\item \href{#method-LorenzSystem-clone}{\code{LorenzSystem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-new"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$new(sigma, R, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-pars"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-pars}{}}}
\subsection{Method \code{pars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$set_params(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-set_state"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-set_state}{}}}
\subsection{Method \code{set_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$set_state(y, time = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-state"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-state}{}}}
\subsection{Method \code{state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-rates"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-rates}{}}}
\subsection{Method \code{rates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$rates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-set_initial_state"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-set_initial_state}{}}}
\subsection{Method \code{set_initial_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$set_initial_state(y, t0 = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-reset"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LorenzSystem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LorenzSystem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LorenzSystem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
