% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-wrappers.R
\name{low_level_wrappers}
\alias{low_level_wrappers}
\alias{Drivers_new}
\alias{Drivers_set_constant}
\alias{Drivers_set_variable}
\alias{Drivers_set_extrapolate}
\alias{Drivers_evaluate}
\alias{Drivers_evaluate_range}
\alias{Drivers_get_names}
\alias{Drivers_clear}
\alias{OdeControl_new}
\alias{OdeControl_get_controls}
\alias{OdeControl_set_controls}
\alias{OdeControl_set_tol_abs}
\alias{OdeControl_set_tol_rel}
\alias{OdeControl_set_a_y}
\alias{OdeControl_set_a_dydt}
\alias{OdeControl_set_step_size_min}
\alias{OdeControl_set_step_size_max}
\alias{OdeControl_set_step_size_initial}
\title{Low-level Drivers and OdeControl wrappers}
\description{
Low-level exported wrappers around native C++ interfaces for driver
configuration and ODE control. Most users should prefer the higher-level
R6 classes [Drivers] and [OdeControl].
}
\keyword{internal}
