\name{odelia-package}
\alias{odelia-package}
\alias{odelia}
\docType{package}
\title{\packageTitle{odelia}}
\description{\packageDescription{odelia}}
\details{The DESCRIPTION file: \packageDESCRIPTION{odelia}}
\section{Package Content}{\packageIndices{odelia}}
\author{\packageAuthor{odelia}}
\section{Maintainer}{\packageMaintainer{odelia}}
\references{http://www.boost.org/}
\keyword{ package }
